refactor(notification-icon): extract Badge component
