+++ src/components/index.js
@@ -26,3 +26,4 @@ export * from './user-profile.component';
 export * from './view-container.component';
 export * from './image-zoom.component';
 export * from './notification-icon.component';
+export * from './badge.component';
